refactor: revert unnecessary changes from previous PR
