+++ tasklist/client/src/Tasks/Task/Variables/JSONEditorModal/index.tsx
@@ -10,7 +10,7 @@ import {useEffect, useRef, useState} from 'react';
 import {useTranslation} from 'react-i18next';
 import {observer} from 'mobx-react-lite';
 import type {editor} from 'monaco-editor';
-import {isValidJSON} from 'modules/utils/jsonUtils';
+import {isValidJSON} from 'modules/utils/isValidJSON';
 import {Modal} from 'modules/components/Modal';
 import {themeStore} from 'modules/stores/theme';
 import Editor from '@monaco-editor/react';
