+++ tasklist/client/src/Tasks/Task/Variables/validators/index.tsx
@@ -8,7 +8,7 @@
 
 import {createVariableFieldName} from '../createVariableFieldName';
 import {getNewVariablePrefix} from '../getVariableFieldName';
-import {isValidJSON} from 'modules/utils/jsonUtils';
+import {isValidJSON} from 'modules/utils/isValidJSON';
 import {promisifyValidator} from './promisifyValidator';
 import type {FormValues} from '../types';
 import get from 'lodash/get';
