+++ tasklist/client/src/modules/custom-filters/customFiltersSchema.ts
@@ -6,7 +6,7 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {isValidJSON} from 'modules/utils/jsonUtils';
+import {isValidJSON} from 'modules/utils/isValidJSON';
 import {z} from 'zod';
 
 const customFiltersSchema = z.object({
